Footer component wasn't in agreed file structure
