+++ src/components/common/FooterBar.js
@@ -1,0 +1,22 @@
+import ReusableNavLink from '@/components/common/ReusableNavLink'
+
+function FooterBar() {
+  const navList = [
+    { text: 'About', href: '/about' },
+    { text: 'Contact', href: '/contact' },
+    { text: 'Terms of use', href: '/terms of use' },
+  ]
+
+  return (
+    <footer>
+      <span>Logo</span>
+      <section className='navItem'>
+        {navList.map(({ text, href }) => (
+          <ReusableNavLink key={href} text={text} href={href} />
+        ))}
+      </section>
+    </footer>
+  )
+}
+
+export default FooterBar
