+++ src/pages/index.js
@@ -1,18 +1,17 @@
-import Head from "next/head";
-import styles from "@/styles/Home.module.css";
-import Listing from "@/components/landing/listing";
-import { listings } from "@/data/mockListings";
-import FooterBar from "@/common/FooterBar";
+import Head from 'next/head'
+import styles from '@/styles/Home.module.css'
+import Listing from '@/components/landing/listing'
+import { listings } from '@/data/mockListings'
+import FooterBar from '@/components/common/FooterBar'
 
 export default function Home() {
-     
   return (
     <>
       <Head>
         <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
+        <meta name='description' content='Generated by create next app' />
+        <meta name='viewport' content='width=device-width, initial-scale=1' />
+        <link rel='icon' href='/favicon.ico' />
       </Head>
       <main className={styles.main}>
         <section className={styles.listings_container}>
@@ -23,5 +22,5 @@ export default function Home() {
       </main>
       <FooterBar />
     </>
-  );
+  )
 }
